requirements.txt
flask
tensorflow
numpy
Pillow
google-generativeai
python-dotenv